FEAT: Implementando Spring Security

+++ src/main/java/com/github/guilhermemonte21/Ecommerce/Application/UseCase/Carrinho/RemoverItemDoCarrinho/RemoverItemDoCarrinho.java
@@ -2,7 +2,9 @@ package com.github.guilhermemonte21.Ecommerce.Application.UseCase.Carrinho.Remov
 
 import com.github.guilhermemonte21.Ecommerce.Application.Exceptions.CarrinhoNotFoundException;
 import com.github.guilhermemonte21.Ecommerce.Application.Gateway.CarrinhoGateway;
-import com.github.guilhermemonte21.Ecommerce.Domain.Model.Entity.Carrinho;
+import com.github.guilhermemonte21.Ecommerce.Application.Gateway.UsuarioAutenticadoGateway;
+import com.github.guilhermemonte21.Ecommerce.Domain.Entity.UsuarioAutenticado;
+import com.github.guilhermemonte21.Ecommerce.Domain.Entity.Carrinho;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,15 +13,21 @@ import java.util.UUID;
 @Service
 public class RemoverItemDoCarrinho implements IRemoverItemDoCarrinho{
     private final CarrinhoGateway gateway;
+    private final UsuarioAutenticadoGateway AuthGateway;
 
-    public RemoverItemDoCarrinho(CarrinhoGateway gateway) {
+    public RemoverItemDoCarrinho(CarrinhoGateway gateway, UsuarioAutenticadoGateway authGateway) {
         this.gateway = gateway;
+        AuthGateway = authGateway;
     }
 
     @Override
     @Transactional
     public void RemoverItem(UUID IdCarrinho, UUID idProduto){
         Carrinho carrinho = gateway.getById(IdCarrinho).orElseThrow(() -> new CarrinhoNotFoundException(IdCarrinho));
+        UsuarioAutenticado user = AuthGateway.get();
+        if(!user.getId().equals(carrinho.getComprador().getId())){
+            throw new RuntimeException("Acesso Negado");
+        }
         gateway.DeleteItem(carrinho,idProduto);
         carrinho.atualizarValorTotal();
         carrinho.AtualizadoAgora();
